Restructure browser-tools adapters and co-locate tests under src/.
Move the AI SDK entry point to src/adapters/ai-sdk/, update the ./ai-sdk
export path, and relocate specs next to the code they cover.
